# controllers/catalogueController.js
import getPool from "../utils/db.js";
import { isPrivilegedRole, normalizePartnerInsurer } from "../middlewares/roleMiddleware.js";
import path from "path";
import fs from "fs";
import multer from "multer";

const planLogoStorage = multer.diskStorage({
  destination: (req, file, cb) => {
    const dir = path.join(process.cwd(), "uploads", "plan-logos");
    fs.mkdirSync(dir, { recursive: true });
    cb(null, dir);
  },
  filename: (req, file, cb) => {
    const ext = path.extname(file.originalname || "").toLowerCase();
    const allowed = [".png", ".jpg", ".jpeg", ".gif", ".svg", ".webp"];
    const safeExt = allowed.includes(ext) ? ext : ".png";
    cb(null, `plan-${req.params.id}-${Date.now()}${safeExt}`);
  }
});

/** Single image for catalogue partner insurer logo (certificate header). Field name: partner_logo */
export const uploadPartnerLogoMiddleware = multer({
  storage: planLogoStorage,
  limits: { fileSize: 2 * 1024 * 1024 },
  fileFilter: (req, file, cb) => {
    const ext = path.extname(file.originalname || "").toLowerCase();
    const ok = [".png", ".jpg", ".jpeg", ".gif", ".svg", ".webp"].includes(ext);
    if (ok) cb(null, true);
    else cb(new Error("Only PNG, JPG, JPEG, GIF, SVG, or WEBP images are allowed"));
  }
}).single("partner_logo");

function publicAssetUrl(req, relativePath) {
  if (!relativePath) return null;
  const p = String(relativePath).trim();
  if (/^https?:\/\//i.test(p)) return p;
  const envBase = (process.env.BASE_URL || "").replace(/\/$/, "");
  if (envBase) return `${envBase}${p.startsWith("/") ? "" : "/"}${p}`;
  const host = req.get("x-forwarded-host") || req.get("host") || "";
  const proto = req.get("x-forwarded-proto") || req.protocol || "https";
  if (!host) return p.startsWith("/") ? p : `/${p}`;
  return `${proto}://${host}${p.startsWith("/") ? "" : "/"}${p}`;
}

const DEFAULT_THEME_COLOR = "#E4590F";
/** Normalize incoming theme color to "#RRGGBB" or "#RRGGBBAA" (fallback to brand orange) */
function sanitizeThemeColor(input) {
  if (input == null) return DEFAULT_THEME_COLOR;
  const s = String(input).trim();
  if (!s) return DEFAULT_THEME_COLOR;
  return /^#([0-9a-fA-F]{6}|[0-9a-fA-F]{8})$/.test(s) ? s.toUpperCase() : DEFAULT_THEME_COLOR;
}

/**
 * Optional short coverage blurb shown in the WhatsApp quote. Left NULL, the chat
 * message derives its own from the plan's `guarantees`, so this is an override
 * for the cases where the derived text reads badly — not a required field.
 */
function coverageSummaryValue(v) {
  if (v === undefined) return undefined;
  if (v === null) return null;
  const s = String(v).trim();
  return s === "" ? null : s.slice(0, 2000);
}

function toBoolFlag(v) {
  if (v === true || v === 1 || v === "1") return 1;
  if (typeof v === "string" && v.toLowerCase() === "true") return 1;
  return 0;
}

// Create catalogue plan
export const createCatalogue = async (req, res) => {
  try {
    const pool = getPool();
    const {
      product_type,
      name,
      coverage,
      pricing_rules,
      flat_price,
      country_of_residence,
      route_type,
      currency,
      theme_color,
      extra_id_fields,
      fixed_duration_premiums,
      partner_insurer,
      whatsapp_enabled,
      coverage_summary_fr,
      coverage_summary_en
    } = req.body;

    const countryValue =
      country_of_residence && typeof country_of_residence === "string" && country_of_residence.trim() !== ""
        ? country_of_residence.trim()
        : null;

    const routeValue =
      route_type && typeof route_type === "string" && route_type.trim() !== "" ? route_type.trim() : null;

    const coverageValue = coverage != null && String(coverage).trim() !== "" ? String(coverage).trim() : null;
    const partnerInsurerValue = normalizePartnerInsurer(partner_insurer);

    const [result] = await pool.query(
      `INSERT INTO catalogue (product_type, name, coverage, pricing_rules, flat_price, country_of_residence, route_type, currency, theme_color, extra_id_fields, fixed_duration_premiums, partner_insurer, whatsapp_enabled, coverage_summary_fr, coverage_summary_en)
            VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
      [
        product_type,
        name,
        coverageValue,
        JSON.stringify(pricing_rules),
        flat_price,
        countryValue,
        routeValue,
        currency || "XOF",
        sanitizeThemeColor(theme_color),
        toBoolFlag(extra_id_fields),
        toBoolFlag(fixed_duration_premiums),
        partnerInsurerValue,
        // Defaults to 0: enabling the WhatsApp module must never silently expose
        // every existing plan to chat customers.
        toBoolFlag(whatsapp_enabled),
        coverageSummaryValue(coverage_summary_fr) ?? null,
        coverageSummaryValue(coverage_summary_en) ?? null
      ]
    );
    res.status(201).json({
      success: true,
      message: "Catalogue plan created successfully",
      id: result.insertId
    });
  } catch (err) {
    console.error(err);
    res.status(500).json({ success: false, message: "Error creating catalogue" });
  }
};

// Get all catalogues (Admin / sub-admin: all plans; insurer_supervisor: insurer plans; Agent: assigned)
export const getCatalogues = async (req, res) => {
  try {
    const pool = getPool();
    let query;
    const params = [];
    if (isPrivilegedRole(req.user.role)) {
      query = "SELECT * FROM catalogue";
    } else if (req.user.role === "insurer_supervisor") {
      const { findUserById } = await import("../models/userModel.js");
      const u = await findUserById(req.user.id);
      const insurer = u?.partner_insurer;
      if (!insurer) {
        return res.json({ success: true, data: [] });
      }
      query = "SELECT * FROM catalogue WHERE partner_insurer = ?";
      params.push(String(insurer).toLowerCase());
    } else {
      query = `SELECT c.* FROM catalogue c
                INNER JOIN user_assigned_plans uap ON uap.catalogue_id = c.id AND uap.user_id = ?
                WHERE c.active = true`;
      params.push(req.user.id);
    }
    const [rows] = await pool.query(query, params);
    const parsedRows = rows.map((row) => {
      if (row.pricing_rules && typeof row.pricing_rules === "string") {
        try {
          row.pricing_rules = JSON.parse(row.pricing_rules);
        } catch (e) {
          console.error("Error parsing pricing_rules:", e);
          row.pricing_rules = null;
        }
      }
      row.theme_color = sanitizeThemeColor(row.theme_color);
      row.extra_id_fields = !!Number(row.extra_id_fields);
      row.fixed_duration_premiums = !!Number(row.fixed_duration_premiums);
      return row;
    });
    res.json({ success: true, data: parsedRows });
  } catch (err) {
    console.error(err);
    res.status(500).json({ success: false, message: "Error fetching catalogues" });
  }
};

// Update catalogue
export const updateCatalogue = async (req, res) => {
  try {
    const pool = getPool();
    const { id } = req.params;
    const {
      product_type,
      name,
      coverage,
      pricing_rules,
      flat_price,
      active,
      country_of_residence,
      route_type,
      currency,
      theme_color,
      extra_id_fields,
      fixed_duration_premiums,
      partner_insurer,
      whatsapp_enabled,
      coverage_summary_fr,
      coverage_summary_en
    } = req.body;

    const countryValue =
      country_of_residence && typeof country_of_residence === "string" && country_of_residence.trim() !== ""
        ? country_of_residence.trim()
        : null;

    const routeValue =
      route_type && typeof route_type === "string" && route_type.trim() !== "" ? route_type.trim() : null;

    const coverageValue = coverage != null && String(coverage).trim() !== "" ? String(coverage).trim() : null;
    const partnerInsurerValue =
      partner_insurer !== undefined ? normalizePartnerInsurer(partner_insurer) : undefined;

    if (partnerInsurerValue !== undefined) {
      await pool.query(
        `UPDATE catalogue SET product_type=?, name=?, coverage=?, pricing_rules=?, flat_price=?, active=?, country_of_residence=?, route_type=?, currency=?, theme_color=?, extra_id_fields=?, fixed_duration_premiums=?, partner_insurer=?, whatsapp_enabled=?, coverage_summary_fr=COALESCE(?, coverage_summary_fr), coverage_summary_en=COALESCE(?, coverage_summary_en) WHERE id=?`,
        [
          product_type,
          name,
          coverageValue,
          JSON.stringify(pricing_rules),
          flat_price,
          active,
          countryValue,
          routeValue,
          currency || "XOF",
          sanitizeThemeColor(theme_color),
          toBoolFlag(extra_id_fields),
          toBoolFlag(fixed_duration_premiums),
          partnerInsurerValue,
          toBoolFlag(whatsapp_enabled),
          // COALESCE in the SQL: a payload that omits these keeps what is stored,
          // so an edit from a form that does not render them cannot wipe them.
          coverageSummaryValue(coverage_summary_fr) ?? null,
          coverageSummaryValue(coverage_summary_en) ?? null,
          id
        ]
      );
    } else {
      await pool.query(
        `UPDATE catalogue SET product_type=?, name=?, coverage=?, pricing_rules=?, flat_price=?, active=?, country_of_residence=?, route_type=?, currency=?, theme_color=?, extra_id_fields=?, fixed_duration_premiums=?, whatsapp_enabled=?, coverage_summary_fr=COALESCE(?, coverage_summary_fr), coverage_summary_en=COALESCE(?, coverage_summary_en) WHERE id=?`,
        [
          product_type,
          name,
          coverageValue,
          JSON.stringify(pricing_rules),
          flat_price,
          active,
          countryValue,
          routeValue,
          currency || "XOF",
          sanitizeThemeColor(theme_color),
          toBoolFlag(extra_id_fields),
          toBoolFlag(fixed_duration_premiums),
          toBoolFlag(whatsapp_enabled),
          coverageSummaryValue(coverage_summary_fr) ?? null,
          coverageSummaryValue(coverage_summary_en) ?? null,
          id
        ]
      );
    }
    res.json({ success: true, message: "Catalogue updated successfully" });
  } catch (err) {
    console.error(err);
    res.status(500).json({ success: false, message: "Error updating catalogue" });
  }
};

// Delete catalogue
export const deleteCatalogue = async (req, res) => {
  try {
    const pool = getPool();
    const { id } = req.params;
    await pool.query("DELETE FROM catalogue WHERE id = ?", [id]);
    res.json({ success: true, message: "Catalogue deleted successfully" });
  } catch (err) {
    console.error(err);
    res.status(500).json({ success: false, message: "Error deleting catalogue" });
  }
};

/** POST multipart partner_logo — saves under /uploads/plan-logos/ and updates catalogue.partner_insurer_logo */
export const uploadPartnerInsurerLogo = async (req, res) => {
  try {
    if (!req.file) {
      return res.status(400).json({ success: false, message: "No image file (field name: partner_logo)" });
    }
    const { id } = req.params;
    const pool = getPool();
    const [rows] = await pool.query("SELECT partner_insurer_logo FROM catalogue WHERE id = ?", [id]);
    if (!rows.length) {
      try {
        fs.unlinkSync(req.file.path);
      } catch (_) {}
      return res.status(404).json({ success: false, message: "Plan not found" });
    }
    const prev = rows[0].partner_insurer_logo;
    if (prev && typeof prev === "string" && prev.includes("plan-logos")) {
      const oldPath = path.join(process.cwd(), prev.replace(/^\//, ""));
      try {
        if (fs.existsSync(oldPath)) fs.unlinkSync(oldPath);
      } catch (_) {}
    }
    const relative = `/uploads/plan-logos/${path.basename(req.file.path)}`;
    await pool.query("UPDATE catalogue SET partner_insurer_logo = ? WHERE id = ?", [relative, id]);
    res.json({
      success: true,
      partner_insurer_logo: relative,
      partnerLogoUrl: publicAssetUrl(req, relative)
    });
  } catch (err) {
    console.error(err);
    if (req.file?.path) {
      try {
        fs.unlinkSync(req.file.path);
      } catch (_) {}
    }
    res.status(500).json({ success: false, message: "Error uploading partner logo" });
  }
};

/** Remove partner logo file and clear DB field */
export const deletePartnerInsurerLogo = async (req, res) => {
  try {
    const { id } = req.params;
    const pool = getPool();
    const [rows] = await pool.query("SELECT partner_insurer_logo FROM catalogue WHERE id = ?", [id]);
    if (!rows.length) {
      return res.status(404).json({ success: false, message: "Plan not found" });
    }
    const prev = rows[0].partner_insurer_logo;
    if (prev && typeof prev === "string" && prev.includes("plan-logos")) {
      const oldPath = path.join(process.cwd(), prev.replace(/^\//, ""));
      try {
        if (fs.existsSync(oldPath)) fs.unlinkSync(oldPath);
      } catch (_) {}
    }
    await pool.query("UPDATE catalogue SET partner_insurer_logo = NULL WHERE id = ?", [id]);
    res.json({ success: true, message: "Partner logo removed" });
  } catch (err) {
    console.error(err);
    res.status(500).json({ success: false, message: "Error removing partner logo" });
  }
};