# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AssurAssistance Public API
  version: "1.0.0"
  summary: Travel insurance issuance, certificates, invoices and reporting.
  description: |
    Programmatic access to AssurAssistance for travel agencies, partner
    insurers, and field-sales tooling. Use this API to:

      * List the travel-insurance plans available to sell.
      * Create travellers and travel cases, then confirm a sale.
      * Download invoice and certificate PDFs.
      * Mark invoices as paid and reconcile.
      * Pull ledger and "invoices by region" reports.
      * Download contractual documentation (Terms & Conditions full / brief);
        Admin JWT uploads and replaces those files.
      * Admin JWT may change a supervisor / agency login email
        (`PATCH /admin/agents/{id}` + `GET /admin/email-available`); the address
        must be unique among all login accounts.

    The same JSON shapes are used by the in-house AssurAssistance web app, so
    anything the web app can do, a 3rd-party integration can do — within the
    permission scopes granted to its API key.

  contact:
    name: AssurAssistance Integration Support
    email: support@assurassistancepro.org
  license:
    name: Proprietary
servers:
  - url: https://backend-api.assurassistancepro.org/api
    description: Production
  - url: http://localhost:3000/api
    description: Local development
...
security:
  - bearerJWT: []
  - bearerApiKey: []
components:
  securitySchemes:
    bearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by `POST /auth/login`. **Used by the in-house web app only.**
        3rd-party integrations should use the `bearerApiKey` scheme.
    bearerApiKey:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        API key issued by an AssurAssistance admin via `POST /admin/api-keys`.
        Keys look like `aas_live_…`. Pass them as:

        ```
        Authorization: Bearer aas_live_xxxxxxxxxxxxxxxxxxxxxxxxxxxx
        ```

        Each request must satisfy the scopes the key was issued with (see the
        `scopes` table in the Integration Guide).
...
  schemas:
...
    Error:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          properties:
            code: { type: string, example: invalid_token }
            message: { type: string, example: Invalid or expired token }
            requiredScope: { type: string, nullable: true }
            availableScopes:
              type: array
              items: { type: string }
...
    Plan:
      type: object
      description: A travel-insurance plan in the catalogue.
      properties:
        id: { type: integer }
        product_type:
          type: string
          enum: [Travel, "Travel Inbound", Bank, "Health Evacuation", "Road travel"]
        name: { type: string }
        coverage: { type: string, nullable: true }
        eligible_destinations: { type: string, nullable: true }
        durations: { type: string, nullable: true }
        pricing_rules: { type: object, nullable: true }
        flat_price: { type: number, nullable: true }
        terms: { type: string, nullable: true }
        country_of_residence: { type: string, nullable: true }
        route_type: { type: string, nullable: true }
        currency: { type: string, default: XOF }
        theme_color: { type: string, example: "#E4590F" }
        extra_id_fields: { type: boolean }
        partner_insurer_logo: { type: string, nullable: true }
        partner_insurer:
          type: string
          nullable: true
          description: Owning insurer key used to scope insurer supervisors (e.g. gna, agico).
        status:
          type: string
          enum: [active, inactive]
    CreatePlanRequest:
      type: object
      required: [product_type, name]
      properties:
        product_type:
          type: string
          enum: [Travel, "Travel Inbound", Bank, "Health Evacuation", "Road travel"]
        name: { type: string }
        coverage: { type: string, nullable: true }
        pricing_rules: { type: object, nullable: true }
        flat_price: { type: number, nullable: true }
        country_of_residence: { type: string, nullable: true }
        route_type: { type: string, nullable: true }
        currency: { type: string, default: XOF }
        theme_color: { type: string, example: "#1E4D8B" }
        extra_id_fields: { type: boolean }
...
paths:
...
  /catalogue:
    get:
      tags: [Catalogue]
      summary: List available plans.
      security:
        - bearerJWT: []
        - bearerApiKey: [catalogue:read]
      responses:
        "200":
          description: Plan list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Plan' }
    post:
      tags: [Catalogue]
      summary: Create a plan (admin only).
      description: Internal admin endpoint. No API-key scope grants access.
      security:
        - bearerJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePlanRequest' }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
                  id: { type: integer }
        "500":
          description: Server error.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /catalogue/{id}:
    put:
      tags: [Catalogue]
      summary: Update a plan (admin only).
      security:
        - bearerJWT: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePlanRequest' }
      responses:
        "200": { description: Updated. }
    delete:
      tags: [Catalogue]
      summary: Delete a plan (admin only).
      security:
        - bearerJWT: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Deleted. }
...